requirements.txt
Django==1.11.14
django-adminlte==1.2
django-adminlte2==0.1.6
django-appconf==1.0.2
django-datetime-widget-fa==0.10.0
Django-Select2==5.11.1
mysqlclient==1.3.9
Pillow==5.2.0
pytz==2018.5
reportlab==3.5.6
